Basic Idea and Layout
There is a basic UI with a header and some placeholder text. I also made a couple of buttons that have functions and disappear after being clicked. The attack function is almost finished and the defense just adds text to the screen and it need to be completed.  There is a character object that has the hero status. There is a constructor function that make monsters but right now it is only used once.