Attack and Defend Method
I created a constructor function for locations and made a few. I added a few more monsters. Attack and defend are now methods for yourCharacter and they take an argument so the monster can change when called. I also changed a few vars to let. The buttons display is equal to none after they are clicked.